+++ src/components/article/ArticleBox/ArticleBox.tsx
@@ -1,8 +1,8 @@
-import { Article as ArticleProps } from "src/models";
+import { Article as ArticleProps } from "src/models/article";
 import styled from "@emotion/styled";
 import Link from "next/link";
 import Image from "next/image";
-import { TagList } from "src/components/commons";
+import { Chip } from "src/components/commons";
 
 interface Props {
   article: ArticleProps;
@@ -14,13 +14,23 @@ const ArticleBox: React.FC<Props> = ({ article }: Props) => {
     <>
       <Link href={`/${email}/${article._id}`}>
         <Container>
-          <ImageContainer>
-            <Image src={thumbnailUrl} fill alt={title} />
-          </ImageContainer>
-          <Title>{title}</Title>
-          <Introduction>{introduction}</Introduction>
-          <TagList tags={tags} size="mini" />
-          <Time>{syncTime}</Time>
+          {thumbnailUrl && (
+            <ImageContainer>
+              <Image src={thumbnailUrl} fill alt={title} />
+            </ImageContainer>
+          )}
+          <Title data-testid="article-title">{title}</Title>
+          <Introduction data-testid="article-description">
+            {introduction}
+          </Introduction>
+          <TagContainer>
+            {tags.map((tag: string) => (
+              <Chip size="mini" key={tag} data-testid="article-item-tag">
+                {tag}
+              </Chip>
+            ))}
+          </TagContainer>
+          <Time data-testid="article-time">{syncTime}</Time>
         </Container>
       </Link>
     </>
@@ -29,11 +39,17 @@ const ArticleBox: React.FC<Props> = ({ article }: Props) => {
 
 const Container = styled.div`
   width: 650px;
-  height: 450px;
+  height: 100%;
   display: flex;
   flex-direction: column;
   cursor: pointer;
   gap: 0.6rem;
+`;
+
+const TagContainer = styled.ul`
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.7rem;
 `;
 
 const Title = styled.h1`
@@ -50,8 +66,8 @@ const Time = styled.h3`
 `;
 
 const ImageContainer = styled.div`
-  width: 100%;
-  height: 100%;
+  width: 768px;
+  height: 400px;
   position: relative;
 `;
 export default ArticleBox;
