+++ src/components/commons/Chip/index.ts
@@ -1,1 +1,1 @@
-export { default as TagList } from "./TagList";
+export { default as Chip } from "./Chip";
